TestUnitedByUnited/TESTES/sw-planet-api/src/main/java/com/example/swplanetapi/service/PlanetService.java
package com.example.Swplanetapi.service;

import com.example.Swplanetapi.domain.Planet;
import com.example.Swplanetapi.domain.QueryBuilder;
import com.example.Swplanetapi.respository.PlanetRepository;
import org.springframework.data.domain.Example;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class PlanetService {
    private PlanetRepository planetRepository;

    public PlanetService(PlanetRepository planetRepository){
        this.planetRepository = planetRepository;
    }
    public Planet create(Planet planet){
        return planetRepository.save(planet);
    }

    public Optional<Planet> get(Long id) {
        return planetRepository.findById(id);
    }

    public Optional<Planet> getByName(String name) {
        return planetRepository.findByName(name);
    }

    public List<Planet> list(String name, String terrain, String climate) {
        Example<Planet> query = QueryBuilder.makeQuery(new Planet(name ,climate));
        return planetRepository.findAll(query);
    }

    public void remove(Long id) {
        planetRepository.deleteById(id);
    }
}
